refactor: Separate GUI from core logic into MVC pattern
This commit refactors the entire application to separate the user interface from the core business logic, following a Model-View-Controller (MVC) architecture.

- A new `iso_logic.py` file has been created to act as the Model. It contains:
  - `ISOCore`: A new class that encapsulates all the state and methods for loading, parsing, modifying, and saving ISO files.
  - `ISOBuilder`: The complex ISO building class and its helpers have been moved here.

- The `ISO_edit.py` file has been significantly simplified and now acts as the View/Controller:
  - The `ISOEditor` class no longer contains any direct ISO manipulation logic.
  - It now holds an instance of `ISOCore` and delegates all logic calls to it.
  - Its sole responsibility is to manage the `tkinter` GUI, handle user input, and display data provided by the `ISOCore` model.

This refactoring makes the codebase much more maintainable, modular, and, most importantly, allows for the core logic to be tested independently of the GUI.
